# Make ByteStreamServerProxy call ByteStreamServer directly