This changes the ByteStreamServerProxy to use the `Read()`, `BeginWrite()`, and `WriteState.Write()` functions directly, the latter two of which were introduced in https://github.com/buildbuddy-io/buildbuddy/pull/9105.

Benchmark results:
```
goos: linux
goarch: amd64
cpu: AMD Ryzen 7 5800X 8-Core Processor             
         │ /home/iain/old.txt │         /home/iain/new.txt          │
         │       sec/op       │   sec/op     vs base                │
Read-16          404.2µ ±  7%   226.6µ ± 5%  -43.94% (p=0.000 n=25)
Write-16         665.7µ ± 10%   477.9µ ± 6%  -28.22% (p=0.000 n=25)
geomean          518.8µ         329.1µ       -36.56%

         │ /home/iain/old.txt │          /home/iain/new.txt          │
         │        B/op        │     B/op      vs base                │
Read-16          149.2Ki ± 0%   104.7Ki ± 0%  -29.78% (p=0.000 n=25)
Write-16         175.7Ki ± 0%   130.8Ki ± 0%  -25.54% (p=0.000 n=25)
geomean          161.9Ki        117.0Ki       -27.69%

         │ /home/iain/old.txt │         /home/iain/new.txt         │
         │     allocs/op      │ allocs/op   vs base                │
Read-16            782.0 ± 0%   439.0 ± 0%  -43.86% (p=0.000 n=25)
Write-16          1267.0 ± 0%   928.0 ± 0%  -26.76% (p=0.000 n=25)
geomean            995.4        638.3       -35.88%
```

Related issues: https://github.com/buildbuddy-io/buildbuddy-internal/issues/4782