@iain-macdonald 2025-05-01
The diff here is kinda gnarly (sorry), but IMO the resulting code in `byte_stream_server_proxy.go` is a lot easier to read. I thought about trying to split this into two changes: one to restructure, and one to swap out the ByteStreamClient for a ByteStreamServer, but that's a bit of work and I feel like this is manageable. LMK if you disagree though.

https://github.com/buildbuddy-io/buildbuddy/pull/9160 picks off some of the diff from here.