@iain-macdonald 2025-05-01
> Did any of the changes due to PR comments change the benchmark performance?

Doesn't look like it. Guessing that there's enough other stuff going on here that this improvements don't quite make their way into the benchmark results.:
```
goos: linux
goarch: amd64
cpu: AMD Ryzen 7 5800X 8-Core Processor             
         │ /home/iain/pre-vanja.txt │     /home/iain/post-vanja.txt      │
         │          sec/op          │   sec/op     vs base               │
Read-16                 228.5µ ± 2%   231.5µ ± 1%  +1.30% (p=0.037 n=25)
Write-16                462.0µ ± 2%   463.9µ ± 2%       ~ (p=0.985 n=25)
geomean                 324.9µ        327.7µ       +0.85%

         │ /home/iain/pre-vanja.txt │      /home/iain/post-vanja.txt      │
         │           B/op           │     B/op      vs base               │
Read-16                107.3Ki ± 0%   107.3Ki ± 0%       ~ (p=0.747 n=25)
Write-16               133.4Ki ± 0%   133.5Ki ± 0%  +0.13% (p=0.000 n=25)
geomean                119.6Ki        119.7Ki       +0.06%

         │ /home/iain/pre-vanja.txt │     /home/iain/post-vanja.txt     │
         │        allocs/op         │ allocs/op   vs base               │
Read-16                  438.0 ± 0%   436.0 ± 0%  -0.46% (p=0.000 n=25)
Write-16                 930.0 ± 0%   931.0 ± 0%  +0.11% (p=0.000 n=25)
geomean                  638.2        637.1       -0.17%
```